+++ automation/src/epilot/models/components/cartcheckoutaction.py
@@ -2,10 +2,12 @@
 
 from __future__ import annotations
 import dataclasses
+from .cartcheckoutconfig import CartCheckoutConfig
 from .erroroutput import ErrorOutput
 from .executionstatus import ExecutionStatus
 from .retrystrategy import RetryStrategy
 from dataclasses_json import Undefined, dataclass_json
+from enum import Enum
 from epilot import utils
 from typing import Any, Dict, Optional
 
@@ -20,6 +22,9 @@ class CartCheckoutActionReason:
     
 
 
+class CartCheckoutActionType(str, Enum):
+    CART_CHECKOUT = 'cart-checkout'
+
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
 @dataclasses.dataclass
@@ -27,7 +32,7 @@ class CartCheckoutAction:
     r"""Creates an order entity with prices from journey"""
     allow_failure: Optional[bool] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('allow_failure'), 'exclude': lambda f: f is None }})
     r"""Whether to stop execution in a failed state if this action fails"""
-    config: Optional[Dict[str, Any]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('config'), 'exclude': lambda f: f is None }})
+    config: Optional[CartCheckoutConfig] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('config'), 'exclude': lambda f: f is None }})
     created_automatically: Optional[bool] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('created_automatically'), 'exclude': lambda f: f is None }})
     r"""Flag indicating whether the action was created automatically or manually"""
     error_output: Optional[ErrorOutput] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('error_output'), 'exclude': lambda f: f is None }})
@@ -40,7 +45,7 @@ class CartCheckoutAction:
     retry_strategy: Optional[RetryStrategy] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('retry_strategy'), 'exclude': lambda f: f is None }})
     r"""different behaviors for retrying failed execution actions."""
     started_at: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('started_at'), 'exclude': lambda f: f is None }})
-    type: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type'), 'exclude': lambda f: f is None }})
+    type: Optional[CartCheckoutActionType] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type'), 'exclude': lambda f: f is None }})
     updated_at: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('updated_at'), 'exclude': lambda f: f is None }})